requirements.txt
Flask~=2.2.2
requests~=2.28.1
pymongo~=4.3.3
Django~=4.1.4
openpyxl~=3.0.10
pytest~=7.2.0
pytest-django~=4.5.2
pandas~=1.5.2
uuid~=1.30
numpy~=1.24.1
djangorestframework~=3.14.0